requirements.txt
# ── Core ──────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
sqlmodel>=0.0.19
celery[redis]>=5.4.0
flower>=2.0.1
redis>=5.0.0
httpx>=0.27.0
python-multipart>=0.0.9
websockets>=13.0

# ── PyTorch (CUDA 12.x for A100) ─────────────────────────────────
# Install separately: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
torch>=2.4.0
torchvision>=0.19.0

# ── Diffusers / HuggingFace ──────────────────────────────────────
diffusers>=0.33.0          # 0.33.0+ required for TeaCache hooks
transformers>=4.45.0
accelerate>=0.34.0
huggingface_hub>=0.25.0
safetensors>=0.4.4

# ── Flash Attention (A100 native) ────────────────────────────────
# Install separately: pip install flash-attn --no-build-isolation

# ── SageAttention (1.5-1.7× faster than FA2, drop-in) ───────────
# Install separately on GPU nodes: pip install sageattention
# sageattention>=1.0.0

# ── Quantization (INT8/NF4 for Hunyuan on A100) ─────────────────
bitsandbytes>=0.43.0

# ── Multi-GPU sequence parallelism (xDiT) ────────────────────────
# Install separately on multi-GPU nodes: pip install xdit
# xdit>=0.4.0

# ── Audio ────────────────────────────────────────────────────────
kokoro>=0.9.4
soundfile>=0.12.1
numpy>=1.26.0

# ── LLM / Prompts ────────────────────────────────────────────────
anthropic>=0.34.0

# ── Config ───────────────────────────────────────────────────────
pyyaml>=6.0.2
python-dotenv>=1.0.1

# ── Video export (required by diffusers export_to_video) ─────────
imageio>=2.34.0
imageio-ffmpeg>=0.5.0
opencv-python-headless>=4.9.0
